Assemble phases for -Vphases
